Add corrections

+++ backend/src/main/java/com/onlinebookstore/service/BookServiceImpl.java
@@ -1,5 +1,7 @@
 package com.onlinebookstore.service;
 
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+
 import com.onlinebookstore.model.Book;
 import com.onlinebookstore.model.SearchParameters;
 import com.onlinebookstore.repository.book.BookRepository;
@@ -9,6 +11,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
@@ -34,7 +37,10 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
-        bookRepository.deleteById(id);
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+        }
+        throw new ResponseStatusException(NOT_FOUND, "Book not found with id: " + id);
     }
 
     @Override
@@ -47,11 +53,6 @@ public class BookServiceImpl implements BookService {
         book.setDescription(updatedBook.getDescription());
         book.setCoverImage(updatedBook.getCoverImage());
         return bookRepository.save(book);
-    }
-
-    @Override
-    public boolean existsById(Long id) {
-        return bookRepository.existsById(id);
     }
 
     @Override
